Kanban board can be filtered by story target.
We do not use backbone.couchdb.js for the target
selection widget because our view -> model mapping
is outside of that supported by couchdb-backbone.
It can be made to work by dropping some custom
view handling code but then we aren't really using
couchdb-backbone anymore.

In this case it makes more sense just to override
Backbone.sync with the implementation we require.