Story target widgets are not permanently bound to the target collection. We bind only when needed (just before we add a new target) and unbind straight after.
If we don't do this the target view never goes out
of scope (as we only have a single collection which
persists the lifetime of the session) and we end up
with a target view for every story dialog created
which is responding to events on the single target
collection. This would be bad.